requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
requests>=2.32.0
python-dotenv>=1.0.1
pyjwt>=2.10.0
passlib[bcrypt]>=1.7.4
email-validator>=2.2.0
pydantic[email]>=2.10.0